Resolve GPDB_12_MERGE_FIXMEs: TOAST for AO_ROW Table
This is intended to keep TOAST enabled for AO_ROW table.
Although large contents of attributes can be fully stored
in AO_ROW table, it is still necessary to keep TOAST capability
for AO_ROW to deal with out-of-line data storage case.
Queries (like select pg_relation_size()) will get better
performance for skipping fetching the out-of-line data unless
it is required. It is a good way of keeping tables being stored
locally and compactly.

However, for AOCS tables, we removed TOAST long time ago, because
there is meaningless to store a column data outside the column
storage itself. Long length content for a single column is not
a problem and no row-like overflow situation in column oriented
storage. Queries on AOCS tables always return all data without
differentiating in/out line mode. So TOAST is not used for AOCS
or column orientated table naturally by design.