Fix flaky fts_unblock_primary isolation2 test
The fts_unblock_primary isolation2 test would sometimes fail stating
that the GUC value was not updated after running `gpstop -u` to reload
from the updated postgresql.conf file. The cause of the flakiness was
due to the assumption that the SHOW command would immediately be
usable after `gpstop -u` returned. The `gpstop -u` command merely
sends SIGHUP signals to all the segments (e.g. pg_ctl reload on all
the segments) and does not do any validations. In this test, the SHOW
command was being run before the SIGHUP signal was handled which
resulted in the incorrect value being displayed. To fix this test
flake, we simply need to close our SQL connection after sending the
SIGHUP signals and open a new connection to run the SHOW command. This
solution works because the new connection will block until the config
reload has completed.